+++ src/medfate_runners.cpp
@@ -198,12 +198,12 @@ GROWTH_runner::GROWTH_runner(Rcpp::List x_list,
     if(x_m.control.transpirationMode=="Granier") {
       BasicTranspiration_RESULT BTres(numCohorts, nlayers);
       BasicSPWB_RESULT BSPWBres(BTres);
-      BasicGROWTH_RESULT BGROWTHres(BSPWBres, numCohorts);
+      BasicGROWTH_RESULT BGROWTHres(&BSPWBres, numCohorts);
       GROWTHres = std::make_unique<BasicGROWTH_RESULT>(BGROWTHres);
     } else {
       AdvancedTranspiration_RESULT ATres(numCohorts, nlayers, ncanlayers, ntimesteps);
       AdvancedSPWB_RESULT ASPWBres(ATres);
-      AdvancedGROWTH_RESULT AGROWTHres(ASPWBres, numCohorts, ntimesteps);
+      AdvancedGROWTH_RESULT AGROWTHres(&ASPWBres, numCohorts, ntimesteps);
       GROWTHres = std::make_unique<AdvancedGROWTH_RESULT>(AGROWTHres);
     }
     GROWTHcomm = GROWTHCommunicationStructures(x_m.cohorts.SpeciesIndex.size(), x_m.soil.getNlayers(), x_m.canopy.zlow.size(), x_m.control.advancedWB.ndailysteps);
@@ -259,12 +259,12 @@ GROWTH_multiple_runner::GROWTH_multiple_runner(List wbInput_vec,
     if(x_i.control.transpirationMode=="Granier") {
       BasicTranspiration_RESULT BTres(numCohorts_i, nlayers_i);
       BasicSPWB_RESULT BSPWBres(BTres);
-      BasicGROWTH_RESULT BGROWTHres(BSPWBres, numCohorts_i);
+      BasicGROWTH_RESULT BGROWTHres(&BSPWBres, numCohorts_i);
       GROWTHres_vec[i] = std::make_unique<BasicGROWTH_RESULT>(BGROWTHres);
     } else {
       AdvancedTranspiration_RESULT ATres(numCohorts_i, nlayers_i, ncanlayers_i, ntimesteps_i);
       AdvancedSPWB_RESULT ASPWBres(ATres);
-      AdvancedGROWTH_RESULT AGROWTHres(ASPWBres, numCohorts_i, ntimesteps_i);
+      AdvancedGROWTH_RESULT AGROWTHres(&ASPWBres, numCohorts_i, ntimesteps_i);
       GROWTHres_vec[i] = std::make_unique<AdvancedGROWTH_RESULT>(AGROWTHres);
     }
     numCohorts_max = std::max(numCohorts_max, x_i.cohorts.SpeciesIndex.size());