+++ src/numerical_solving_c.h
@@ -7,4 +7,6 @@ void tridiagonalSolving_c(const std::vector<double> &a, const std::vector<double
                           std::vector<double> &e, std::vector<double> &f, std::vector<double> &sol);
 
 double normal_cdf(double x, double mean = 0.0, double stddev = 1.0);
+double vecsum(const std::vector<double>& vec);
+
 #endif